+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "Ycommunity",
-    "slug": "Ycommunity",
+    "name": "test",
+    "slug": "test",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
